requirements.txt
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
pathlib>=1.0.1
jinja2>=3.1.2
networkx>=3.1
matplotlib>=3.7.1
pandas>=2.0.0
